CBP-21 - Typescript Contact (#7)
* adds: typescript support to Testimonials file

* adds: testimonial images

* adds: explicit type annotations

---------
